requirements.txt
requests>=2.28
python-dotenv>=0.21
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
google-api-python-client>=2.80
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
